+++ TacticalAI/AIMain.cpp
@@ -1137,7 +1137,7 @@ BOOLEAN DestNotSpokenFor(SOLDIERTYPE *pSoldier, INT32 sGridNo)
 INT32 FindAdjacentSpotBeside(SOLDIERTYPE *pSoldier, INT32 sGridNo)
 {
 	INT32 cnt;
-	INT16 mods[4] = {-1,-WORLD_COLS,1,WORLD_COLS};
+	INT16 mods[4] = {-1,static_cast<INT16>(-WORLD_COLS),1,static_cast<INT16>(WORLD_COLS)};
 	INT32 sTempGridNo = NOWHERE, sCheapestDest = MAX_MAP_POS;
 	INT16 sCheapestCost=500, sMovementCost;
